API

API

SPICE.SpiceCharCell — Method.

SpiceCharCell(size, length)

Create a SpiceCharCell that can contain up to size strings with length characters.

source

SPICE.SpiceDoubleCell — Method.

SpiceDoubleCell(size)

Create a SpiceDoubleCell that can contain up to size elements.

source

SPICE.SpiceIntCell — Method.

SpiceIntCell(size)

Create a SpiceIntCell that can contain up to size elements.

source

Base.append! — Method.

append!(cell, collection)

Append all items from collection to the char/double/integer SpiceCell cell.

source

Base.length — Method.

length(cell)

source

Base.push! — Method.

push!(cell, items...)

Insert one or more items at the end of the char/double/integer SpiceCell cell.

source

SPICE.appnd — Method.

appnd(item, cell)

Append an item to the char/double/integer SpiceCell cell.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/appndc_c.html https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/appndd_c.html https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/appndi_c.html

source

SPICE.axisar — Method.

axisar(axis, angle)

Construct a rotation matrix that rotates vectors by a specified angle about a specified axis.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/axisar_c.html

source

SPICE.b1900 — Method.

b1900()

Returns the Julian Date corresponding to Besselian date 1900.0.

References

NAIF Documentation

source

SPICE.b1950 — Method.

b1950()

Returns the Julian Date corresponding to Besselian date 1950.0.

References

NAIF Documentation

source

SPICE.badkpv — Method.

badkpv(caller, name, comp, size, divby, typ)

Determine if a kernel pool variable is present and if so that it has the correct size and type.

Arguments

caller: Name of the routine calling this routine

name: Name of a kernel pool variable

comp: Comparison operator

size: Expected size of the kernel pool variable

divby: A divisor of the size of the kernel pool variable

type: Expected type of the kernel pool variable

Output

The function returns false if the kernel pool variable is OK otherwise and exception is thrown.

References

NAIF Documentation

source

SPICE.bltfrm — Method.

bltfrm(frmcls)

Return a SPICE set containing the frame IDs of all built-in frames of a specified class.

Arguments

frmcls: Frame class

Output

idset: Set of ID codes of frames of the specified class

References

NAIF Documentation

source

SPICE.bodc2n — Method.

bodc2n(code)

Translate the SPICE integer code of a body into a common name for that body.

Arguments

code: Integer ID code to be translated into a name

Output

name: A common name for the body identified by code

References

NAIF Documentation

source

SPICE.bodc2s — Method.

bodc2s(code)

Translate a body ID code to either the corresponding name or if no name to ID code mapping exists, the string representation of the body ID value.

Arguments

code: Integer ID code to translate to a string

Output

name: String corresponding to code

References

NAIF Documentation

source

SPICE.boddef — Method.

boddef(name, code)

Define a body name/ID code pair for later translation via bodn2c or bodc2n.

Arguments

name: Common name of some body

code: Integer code for that body

References

NAIF Documentation

source

SPICE.bodfnd — Method.

bodfnd(body, item)

Determine whether values exist for some item for any body in the kernel pool.

Arguments

body: ID code of body

item: Item to find ("RADII", "NUT_AMP_RA", etc.)

Output

found: true if the item is in the kernel pool and false if it is not.

References

NAIF Documentation

source

SPICE.bodn2c — Method.

bodn2c(name)

Translate the name of a body or object to the corresponding SPICE integer ID code.

Arguments

name: Body name to be translated into a SPICE ID code

Output

code: SPICE integer ID code for the named body

References

NAIF Documentation

source

SPICE.bods2c — Method.

bods2c(name)

Translate a string containing a body name or ID code to an integer code.

Arguments

name: String to be translated to an ID code

Output

code: Integer ID code corresponding to name

References

NAIF Documentation

source

SPICE.bodvcd — Method.

bodvcd(bodyid, item)

Fetch from the kernel pool the double precision values of an item associated with a body, where the body is specified by an integer ID code.

Arguments

bodyid: Body ID code

item: Item for which values are desired. ("RADII", "NUT_PREC_ANGLES", etc.)

Output

values: Values

References

NAIF Documentation

source

SPICE.bodvrd — Function.

bodvrd(bodynm, item)

Fetch from the kernel pool the double precision values of an item associated with a body.

Arguments

bodynm: Body name

item: Item for which values are desired. ("RADII", "NUT_PREC_ANGLES", etc.)

maxn: Maximum number of values that may be returned.

Output

values: Values

References

NAIF Documentation

source

SPICE.card — Method.

card(cell)

Returns the cardinality (number of elements) of a SpiceCell cell.

source

SPICE.ccifrm — Method.

ccifrm(frclss, clssid)

Return the frame name, frame ID, and center associated with a given frame class and class ID.

Arguments

frclss: Class of frame

clssid: Class ID of frame

Output

Return the tuple (frcode, frname, center).

frcode: ID code of the frame

frname: Name of the frame

center: ID code of the center of the frame

References

NAIF Documentation

source

SPICE.cidfrm — Method.

cidfrm(cent)

Retrieve frame ID code and name to associate with a frame center.

Arguments

cent: ID code for an object for which there is a preferred reference frame

Output

Returns the tuple (frcode, frname)

frcode: The ID code of the frame associated with cent

frname: The name of the frame with ID frcode

References

NAIF Documentation

source

SPICE.ckcls — Method.

ckcls(handle)

Close an open CK file.

Arguments

handle: Handle of the CK file to be closed

References

NAIF Documentation

source

SPICE.ckgp — Method.

source

SPICE.ckgpav — Method.

source

SPICE.ckopn — Function.

ckopn(fname, ifname="CK_file", ncomch=0)

Open a new CK file, returning the handle of the opened file.

Arguments

fname: The name of the CK file to be opened

ifname="CK_file": The internal filename for the CK, default is "CK_file"

ncomch=0: The number of characters to reserve for comments, default is zero

Output

handle: The handle of the opened CK file

References

NAIF Documentation

source

SPICE.ckw01 — Function.

source

SPICE.clight — Method.

Returns the speed of light in vacuo (km/sec).

source

SPICE.hx2dp — Method.

hx2dp(str)

Convert a string representing a double precision number in a base 16 "scientific notation" into its equivalent double precision number.

Arguments

str: Hex form string to convert to double precision

Output

dp: Double precision value to be returned

References

NAIF Documentation

source

SPICE.j1900 — Method.

j1900()

Returns the Julian Date of 1899 DEC 31 12:00:00 (1900 JAN 0.5).

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/j1900_c.html

source

SPICE.j1950 — Method.

j1950()

Returns the Julian Date of 1950 JAN 01 00:00:00 (1950 JAN 1.0).

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/j1950_c.html

source

SPICE.j2000 — Method.

j2000()

Returns the Julian Date of 2000 JAN 01 12:00:00 (2000 JAN 1.5).

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/j2000_c.html

source

SPICE.j2100 — Method.

j2100()

Returns the Julian Date of 2100 JAN 01 12:00:00 (2100 JAN 1.5).

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/j2100_c.html

source

SPICE.jyear — Method.

jyear()

Returns the number of seconds per Julian year.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/jyear_c.html

source

SPICE.kclear — Method.

kclear()

Clear the KEEPER subsystem: unload all kernels, clear the kernel pool, and re-initialize the subsystem. Existing watches on kernel variables are retained.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/kclear_c.html

source

SPICE.ktotal — Method.

ktotal(kind)

Return the current number of kernels that have been loaded via the KEEPER interface that are of a specified type.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/ktotal_c.html

source

SPICE.pxform — Method.

Return the matrix that transforms position vectors from one specified frame to another at a specified epoch.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/pxform_c.html

source

SPICE.spd — Method.

Returns the number of seconds in a day.

source

SPICE.spkcpo — Method.

Returns the state of a target body relative to a constant-position observer location.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/spkcpo_c.html

source

SPICE.spkezr — Method.

Returns the state of a target body relative to an observing body.

NAIF documentation

source

SPICE.spkpos — Method.

Returns the state of a target body relative to an observing body.

NAIF documentation

source

SPICE.sxform — Method.

Return the state transformation matrix from one frame to another at a specified epoch.

https://naif.jpl.nasa.gov/pub/naif/toolkit_docs/C/cspice/sxform_c.html

source

SPICE.tyear — Method.

Returns the number of seconds per tropical year.

source