API

API

SPICE.SpiceCharCell — Method.

SpiceCharCell(size, length)

Create a SpiceCharCell that can contain up to size strings with length characters.

source

SPICE.SpiceDoubleCell — Method.

SpiceDoubleCell(size)

Create a SpiceDoubleCell that can contain up to size elements.

source

SPICE.SpiceIntCell — Method.

SpiceIntCell(size)

Create a SpiceIntCell that can contain up to size elements.

source

Base.append! — Method.

append!(cell, collection)

Append all items from collection to the char/double/integer SpiceCell cell.

source

Base.copy — Method.

copy(cell::SpiceCell)

Duplicate the SpiceCell cell.

source

Base.length — Method.

length(cell)

Returns the cardinality (number of elements) of a SpiceCell cell.

source

Base.push! — Method.

push!(cell, items...)

Insert one or more items at the end of the char/double/integer SpiceCell cell.

source

Base.union — Method.

union(a::T, b::T) where T <: SpiceCell

Compute the union of two sets of any data type to form a third set.

Arguments

a: First input set

b: Second input set

Output

Returns a cell containing the union of a and b.

References

NAIF Documentation

source

SPICE.appnd — Method.

appnd(item, cell)

Append an item to the char/double/integer SpiceCell cell.

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/appndcc.html https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/appnddc.html https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/appndic.html

source

SPICE.axisar — Method.

axisar(axis, angle)

Construct a rotation matrix that rotates vectors by a specified angle about a specified axis.

Arguments

axis: Rotation axis

angle: Rotation angle in radians

Output

Rotation matrix corresponding to axis and angle

References

NAIF Documentation

source

SPICE.b1900 — Method.

b1900()

Returns the Julian Date corresponding to Besselian date 1900.0.

References

NAIF Documentation

source

SPICE.b1950 — Method.

b1950()

Returns the Julian Date corresponding to Besselian date 1950.0.

References

NAIF Documentation

source

SPICE.badkpv — Method.

badkpv(caller, name, comp, size, divby, typ)

Determine if a kernel pool variable is present and if so that it has the correct size and type.

Arguments

caller: Name of the routine calling this routine

name: Name of a kernel pool variable

comp: Comparison operator

size: Expected size of the kernel pool variable

divby: A divisor of the size of the kernel pool variable

type: Expected type of the kernel pool variable

Output

The function returns false if the kernel pool variable is OK otherwise an exception is thrown.

References

NAIF Documentation

source

SPICE.bltfrm — Method.

bltfrm(frmcls)

Return a SPICE set containing the frame IDs of all built-in frames of a specified class.

Arguments

frmcls: Frame class

Output

idset: Set of ID codes of frames of the specified class

References

NAIF Documentation

source

SPICE.bodc2n — Method.

bodc2n(code)

Translate the SPICE integer code of a body into a common name for that body.

Arguments

code: Integer ID code to be translated into a name

Output

name: A common name for the body identified by code

References

NAIF Documentation

source

SPICE.bodc2s — Method.

bodc2s(code)

Translate a body ID code to either the corresponding name or if no name to ID code mapping exists, the string representation of the body ID value.

Arguments

code: Integer ID code to translate to a string

Output

name: String corresponding to code

References

NAIF Documentation

source

SPICE.boddef — Method.

boddef(name, code)

Define a body name/ID code pair for later translation via bodn2c or bodc2n.

Arguments

name: Common name of some body

code: Integer code for that body

References

NAIF Documentation

source

SPICE.bodfnd — Method.

bodfnd(body, item)

Determine whether values exist for some item for any body in the kernel pool.

Arguments

body: ID code of body

item: Item to find ("RADII", "NUTAMPRA", etc.)

Output

found: true if the item is in the kernel pool and false if it is not.

References

NAIF Documentation

source

SPICE.bodn2c — Method.

bodn2c(name)

Translate the name of a body or object to the corresponding SPICE integer ID code.

Arguments

name: Body name to be translated into a SPICE ID code

Output

code: SPICE integer ID code for the named body

References

NAIF Documentation

source

SPICE.bods2c — Method.

bods2c(name)

Translate a string containing a body name or ID code to an integer code.

Arguments

name: String to be translated to an ID code

Output

code: Integer ID code corresponding to name

References

NAIF Documentation

source

SPICE.bodvcd — Method.

bodvcd(bodyid, item)

Fetch from the kernel pool the double precision values of an item associated with a body, where the body is specified by an integer ID code.

Arguments

bodyid: Body ID code

item: Item for which values are desired. ("RADII", "NUTPRECANGLES", etc.)

Output

values: Values

References

NAIF Documentation

source

SPICE.bodvrd — Function.

bodvrd(bodynm, item)

Fetch from the kernel pool the double precision values of an item associated with a body.

Arguments

bodynm: Body name

item: Item for which values are desired. ("RADII", "NUTPRECANGLES", etc.)

maxn: Maximum number of values that may be returned.

Output

values: Values

References

NAIF Documentation

source

SPICE.brcktd — Function.

brcktd(number, e1, e2)

Deprecated: Use clamp from Julia's standard library instead.

source

SPICE.brckti — Function.

brckti(number, e1, e2)

Deprecated: Use clamp from Julia's standard library instead.

source

SPICE.bschoc — Function.

bschoc(value, array, order)

Deprecated: Use findfirst(array .== value) instead.

source

SPICE.bschoi — Function.

bschoi(value, array, order)

Deprecated: Use findfirst(array .== value) instead.

source

SPICE.bsrchc — Function.

bsrchc(value, array)

Deprecated: Use findfirst(array .== value) instead.

source

SPICE.bsrchd — Function.

bsrchd(value, array)

Deprecated: Use findfirst(array .== value) instead.

source

SPICE.bsrchi — Function.

bsrchi(value, array)

Deprecated: Use findfirst(array .== value) instead.

source

SPICE.card — Method.

card(cell)

Returns the cardinality (number of elements) of a SpiceCell cell.

source

SPICE.ccifrm — Method.

ccifrm(frclss, clssid)

Return the frame name, frame ID, and center associated with a given frame class and class ID.

Arguments

frclss: Class of frame

clssid: Class ID of frame

Output

Return the tuple (frcode, frname, center).

frcode: ID code of the frame

frname: Name of the frame

center: ID code of the center of the frame

References

NAIF Documentation

source

SPICE.cgv2el — Method.

cgv2el(center, vec1, vec2)

Form a CSPICE ellipse from a center vector and two generating vectors.

Arguments

center: Center vector

vec1: Generating vector

vec2: Generating vector

Output

The CSPICE ellipse defined by the input vectors.

References

NAIF Documentation

source

SPICE.cidfrm — Method.

cidfrm(cent)

Retrieve frame ID code and name to associate with a frame center.

Arguments

cent: ID code for an object for which there is a preferred reference frame

Output

Returns the tuple (frcode, frname)

frcode: The ID code of the frame associated with cent

frname: The name of the frame with ID frcode

References

NAIF Documentation

source

SPICE.ckcls — Method.

ckcls(handle)

Close an open CK file.

Arguments

handle: Handle of the CK file to be closed

References

NAIF Documentation

source

SPICE.ckcov! — Method.

ckcov!(ck, idcode, needav, level, tol, timsys, cover)

Find the coverage window for a specified object in a specified CK file.

Arguments

ck: Name of CK file

idcode: ID code of object

needav: Flag indicating whether angular velocity is needed

level: Coverage level: "SEGMENT" OR "INTERVAL"

tol: Tolerance in ticks

timsys: Time system used to represent coverage

cover: Window giving coverage for idcode. Data already present in cover will be combined with coverage found for the object designated by idcode in the file ck.

References

NAIF Documentation

source

SPICE.ckcov — Method.

ckcov(ck, idcode, needav, level, tol, timsys)

Find the coverage window for a specified object in a specified CK file.

Arguments

ck: Name of CK file

idcode: ID code of object

needav: Flag indicating whether angular velocity is needed

level: Coverage level: "SEGMENT" OR "INTERVAL"

tol: Tolerance in ticks

timsys: Time system used to represent coverage

Output

Window giving coverage for idcode

References

NAIF Documentation

source

SPICE.ckgp — Method.

ckgp(inst, sclkdp, tol, ref)

Get pointing (attitude) for a specified spacecraft clock time.

Arguments

inst: NAIF ID of instrument, spacecraft, or structure

sclkdp: Encoded spacecraft clock time

tol: Time tolerance

ref: Reference frame

Outputs

cmat: C-matrix pointing data

clkout: Output encoded spacecraft clock time

found: true when requested pointing is available

References

NAIF Documentation

source

SPICE.ckgpav — Method.

ckgpav(inst, sclkdp, tol, ref)

Get pointing (attitude) and angular velocity for a specified spacecraft clock time.

Arguments

inst: NAIF ID of instrument, spacecraft, or structure

sclkdp: Encoded spacecraft clock time

tol: Time tolerance

ref: Reference frame

Outputs

cmat: C-matrix pointing data

av: Angular velocity vector

clkout: Output encoded spacecraft clock time

found: true when requested pointing is available

References

NAIF Documentation

source

SPICE.cklpf — Method.

cklpf(filename)

Load a CK pointing file for use by the CK readers. Return that file's handle, to be used by other CK routines to refer to the file.

Arguments

filename: Name of the CK file to be loaded

Output

Loaded file's handle

References

NAIF Documentation

source

SPICE.ckobj! — Method.

ckobj!(ck, ids)

Find the set of ID codes of all objects in a specified CK file.

Arguments

ck: Name of CK file

ids: Set of ID codes of objects in CK file. Data already present in ids will be combined with ID code set found for the file ck.

References

NAIF Documentation

source

SPICE.ckobj — Method.

ckobj(ck)

Find the set of ID codes of all objects in a specified CK file.

Arguments

ck: Name of CK file

Output

Set of ID codes of objects in CK file.

References

NAIF Documentation

source

SPICE.ckopn — Function.

ckopn(fname, ifname="CK_file", ncomch=0)

Open a new CK file, returning the handle of the opened file.

Arguments

fname: The name of the CK file to be opened

ifname="CK_file": The internal filename for the CK, default is "CK_file"

ncomch=0: The number of characters to reserve for comments, default is zero

Output

handle: The handle of the opened CK file

References

NAIF Documentation

source

SPICE.ckupf — Method.

ckupf(handle)

Unload a CK pointing file so that it will no longer be searched by the readers.

Arguments

handle: Handle of CK file to be unloaded

References

NAIF Documentation

source

SPICE.ckw01 — Function.

ckw01(handle, inst, ref, segid, sclkdp, quats, avvs=Matrix{SpiceDouble}(0,0);
    begtim=sclkdp[1], endtim=sclkdp[end])

Add a type 1 segment to a C-kernel.

Arguments

handle: Handle of an open CK file

inst: The NAIF instrument ID code

ref: The reference frame of the segment

segid: Segment identifier

sclkdp: Encoded SCLK times

quats: Quaternions representing instrument pointing

avvs: Angular velocity vectors (optional)

begtim: The beginning encoded SCLK of the segment (optional)

endtim: The ending encoded SCLK of the segment (optional)

References

NAIF Documentation

source

SPICE.clight — Method.

clight()

Returns the speed of light in vacuo (km/sec).

source

SPICE.clpool — Method.

clpool()

Remove all variables from the kernel pool. Watches on kernel variables are retained.

References

NAIF Documentation

source

SPICE.cmprss — Method.

cmprss(delim, n, input)

Compress a character string by removing occurrences of more than n consecutive occurrences of a specified character.

Arguments

delim: Delimiter to be compressed

n: Maximum consecutive occurrences of delim

input: Input string

Output

Returns the compressed string.

References

NAIF Documentation

source

SPICE.cnmfrm — Function.

cnmfrm(cname)

Retrieve frame ID code and name to associate with an object.

Arguments

cname: Name of the object to find a frame for

Output

Returns a tuple of the ID code and the name of the frame associated with cname or nothing if no frame is found.

References

NAIF Documentation

source

SPICE.conics — Method.

conics(elts, et)

Determine the state (position, velocity) of an orbiting body from a set of elliptic, hyperbolic, or parabolic orbital elements.

Arguments

elts: Conic elements

et: Input time

Output

Returns the state of orbiting body at et.

References

NAIF Documentation

source

SPICE.convrt — Method.

convrt(x, in, out)

Take a measurement x, the units associated with x, and units to which x should be converted; return y - the value of the measurement in the output units.

Arguments

x: Number representing a measurement in some units

in: The units in which x is measured

out: Desired units for the measurement

Output

Returns the measurement in the desired units.

References

NAIF Documentation

source

SPICE.cpos — Method.

cpos(str, chars, start)

Find the first occurrence in a string of a character belonging to a collection of characters, starting at a specified location, searching forward.

Arguments

str: Any character string

chars: A collection of characters

start: Position to begin looking for one of chars

Output

Returns the index of the first character of str that is one of the characters in string chars. Returns -1 if none of the characters was found.

References

NAIF Documentation

source

SPICE.cposr — Method.

cposr(str, chars, start)

Find the first occurrence in a string of a character belonging to a collection of characters, starting at a specified location, searching in reverse.

Arguments

str: Any character string

chars: A collection of characters

start: Position to begin looking for one of chars

Output

Returns the index of the last character of str that is one of the characters in string chars. Returns -1 if none of the characters was found.

References

NAIF Documentation

source

SPICE.cvpool — Method.

cvpool(agent)

Indicate whether or not any watched kernel variables that have a specified agent on their notification list have been updated.

Arguments

agent: Name of the agent to check for notices

Output

true if variables for agent have been updated.

References

NAIF Documentation

source

SPICE.cyllat — Method.

cyllat(r, lonc, z)

Convert from cylindrical to latitudinal coordinates.

Arguments

r: Distance of point from z axis

lonc: Cylindrical angle of point from XZ plane (radians)

z: Height of point above XY plane

Output

Returns a tuple of radius, longitude (radians), and latitude (radians).

References

NAIF Documentation

source

SPICE.cylrec — Method.

cylrec(r, lon, z)

Convert from cylindrical to rectangular coordinates.

Arguments

r: Distance of the point of interest from z axis

lon: Cylindrical angle (in radians) of the point of interest from XZ plane

z: Height of the point above XY plane

Output

Returns rectangular coordinates of the point of interest.

References

NAIF Documentation

source

SPICE.cylsph — Method.

cylsph(r, lonc, z)

Convert from cylindrical to spherical coordinates.

Arguments

r: Distance of point from z axis

lonc: Angle (radians) of point from XZ plane

z: Height of point above XY plane

Output

Returns a tuple of distance of the point from the origin, polar angle (co-latitude in radians), and azimuthal angle (longitude).

References

NAIF Documentation

source

SPICE.dtpool — Method.

dtpool(name)

Return the data about a kernel pool variable.

Arguments

name: Name of the variable whose value is to be returned

Output

Returns the tuple (n ,vartype).

n: Number of values returned for name

vartype: Type of the variable

:C if the data is character data

:N if the data is numeric

:X if there is no variable name in the pool

References

NAIF Documentation

source

SPICE.et2utc — Method.

et2utc(et, format, prec)

Convert an input time from ephemeris seconds past J2000 to Calendar, Day-of-Year, or Julian Date format, UTC.

Arguments

et: Input epoch, given in ephemeris seconds past J2000

format: Format of output epoch. It may be any of the following:

:C: Calendar format, UTC

:D: Day-of-Year format, UTC

:J: Julian Date format, UTC

:ISOC: ISO Calendar format, UTC

:ISOD: ISO Day-of-Year format, UTC

prec: Digits of precision in fractional seconds or days

Output

Returns an output time string equivalent to the input epoch, in the specified format.

References

NAIF Documentation

source

SPICE.etcal — Function.

etcal(et, lenout=128)

Convert from an ephemeris epoch measured in seconds past the epoch of J2000 to a calendar string format using a formal calendar free of leapseconds.

Arguments

et: Ephemeris time measured in seconds past J2000

lenout: Length of output string (default: 128)

Output

Returns a standard calendar representation of et.

References

NAIF Documentation

source

SPICE.eul2m — Method.

eul2m(angle3, angle2, angle1, axis3, axis2, axis1)

Construct a rotation matrix from a set of Euler angles.

Arguments

angle3, angle2, angle1: Rotation angles about third, second, and first rotation axes (radians)

axis3, axis2, axis1: Axis numbers of third, second, and first rotation axes

Output

A rotation matrix corresponding to the product of the 3 rotations.

References

NAIF Documentation

source

SPICE.gdpool — Method.

gdpool(name; start=1, room=100)

Return the value of a kernel variable from the kernel pool.

Arguments

name: Name of the variable whose value is to be returned

start: Which component to start retrieving for name (default: 1)

room: The largest number of values to return (default: 100)

Output

Returns an array of values if the variable exists or nothing if not.

References

NAIF Documentation

source

SPICE.halfpi — Function.

halfpi()

Deprecated: Use π/2 instead.

source

SPICE.hx2dp — Method.

hx2dp(str)

Convert a string representing a double precision number in a base 16 "scientific notation" into its equivalent double precision number.

Arguments

str: Hex form string to convert to double precision

Output

dp: Double precision value to be returned

References

NAIF Documentation

source

SPICE.j1900 — Method.

j1900()

Returns the Julian Date of 1899 DEC 31 12:00:00 (1900 JAN 0.5).

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/j1900c.html

source

SPICE.j1950 — Method.

j1950()

Returns the Julian Date of 1950 JAN 01 00:00:00 (1950 JAN 1.0).

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/j1950c.html

source

SPICE.j2000 — Method.

j2000()

Returns the Julian Date of 2000 JAN 01 12:00:00 (2000 JAN 1.5).

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/j2000c.html

source

SPICE.j2100 — Method.

j2100()

Returns the Julian Date of 2100 JAN 01 12:00:00 (2100 JAN 1.5).

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/j2100c.html

source

SPICE.jyear — Method.

jyear()

Returns the number of seconds per Julian year.

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/jyearc.html

source

SPICE.kclear — Method.

kclear()

Clear the KEEPER subsystem: unload all kernels, clear the kernel pool, and re-initialize the subsystem. Existing watches on kernel variables are retained.

References

NAIF Documentation

source

SPICE.ktotal — Method.

ktotal(kind)

Return the current number of kernels that have been loaded via the KEEPER interface that are of a specified type.

References

NAIF Documentation

source

SPICE.lastnb — Method.

lastnb(str)

Return the index of the last non-blank character in a character string.

Arguments

str: Input character string

Output

The function returns the one-based index of the last non-blank character in a character string. If the string is entirely blank or is empty, the value 0 is returned.

References

NAIF Documentation

source

SPICE.latcyl — Method.

latcyl(radius, lon, lat)

Convert from latitudinal coordinates to cylindrical coordinates.

Arguments

radius: Distance of a point from the origin

lon: Angle of the point from the XZ plane in radians

lat: Angle of the point from the XY plane in radians

Output

Return the tuple (r, lonc, z).

r: Distance of the point from the z axis

lonc: Angle of the point from the XZ plane in radians. 'lonc' is set equal to 'lon'

z: Height of the point above the XY plane

References

NAIF Documentation

source

SPICE.latrec — Method.

latrec(radius, lon, lat)

Convert from latitudinal coordinates to rectangular coordinates.

Arguments

radius: Distance of a point from the origin

lon: Angle of the point from the XZ plane in radians

lat: Angle of the point from the XY plane in radians

Output

Return the rectangular coordinates vector of the point.

References

NAIF Documentation

source

SPICE.latsph — Method.

latsph(radius, lon, lat)

Convert from latitudinal coordinates to rectangular coordinates.

Arguments

radius: Distance of a point from the origin

lon: Angle of the point from the XZ plane in radians

lat: Angle of the point from the XY plane in radians

Output

Return the tuple (rho, colat, lons).

rho: Distance of the point from the origin

colat: Angle of the point from positive z axis (radians)

lons: Angle of the point from the XZ plane (radians)

References

NAIF Documentation

source

SPICE.latsrf — Method.

latsrf(method, target, et, fixref, npts, lonlat)

Map array of planetocentric longitude/latitude coordinate pairs to surface points on a specified target body.

The surface of the target body may be represented by a triaxial ellipsoid or by topographic data provided by DSK files.

Arguments

method: Computation method

target: Name of target body

et: Epoch in TDB seconds past J2000 TDB

fixref: Body-fixed, body-centered target body frame

npts: Number of coordinate pairs in input array

lonlat: Array of longitude/latitude coordinate pairs

Output

Returns an array of surface points

References

NAIF Documentation

source

SPICE.lcase — Function.

lcase(in)

Deprecated: Use lowercase(in) instead.

source

SPICE.ldpool — Method.

ldpool(kernel)

Load the variables contained in a NAIF ASCII kernel file into the kernel pool.

Arguments

kernel: Name of the kernel file

Output

None

References

NAIF Documentation

source

SPICE.lgrind — Method.

lgrind(xvals, yvals, x)

Evaluate a Lagrange interpolating polynomial for a specified set of coordinate pairs, at a specified abscissa value. Return the value of both polynomial and derivative.

Arguments

xvals: Abscissa values of coordinate pairs

yvals: Ordinate values of coordinate pairs

x: Point at which to interpolate the polynomial

Output

Returns the tuple (p, dp).

p: The value at x of the unique polynomial of degree n-1 that fits the points in the plane defined by xvals and yvals

dp: The derivative at x of the interpolating polynomial described above

References

NAIF Documentation

source

SPICE.limbpt — Method.

limbpt(method, target, et, fixref, abcorr, corloc, obsrvr, refvec, rolstp, ncuts, schstp, soltol, maxn)

Find limb points on a target body. The limb is the set of points of tangency on the target of rays emanating from the observer. The caller specifies half-planes bounded by the observer-target center vector in which to search for limb points.

The surface of the target body may be represented either by a triaxial ellipsoid or by topographic data.

Arguments

method: Computation method

target: Name of target body

et: Epoch in ephemeris seconds past J2000 TDB

fixref: Body-fixed, body-centered target body frame

abcorr: Aberration correction

corloc: Aberration correction locus

obsrvr: Name of observing body

refvec: Reference vector for cutting half-planes

rolstp: Roll angular step for cutting half-planes

ncuts: Number of cutting half-planes

schstp: Angular step size for searching

soltol: Solution convergence tolerance

maxn: Maximum number of entries in output arrays

Output

Returns the tuple (npts, points, epochs, tangts).

npts: Counts of limb points corresponding to cuts

points: Limb points

epochs: Times associated with limb points

tangts: Tangent vectors emanating from the observer

References

NAIF Documentation

source

SPICE.lmpool — Method.

lmpool(cvals)

Load the variables contained in an internal buffer into the kernel pool.

Arguments

cvals: An array that contains a SPICE text kernel

Output

None

References

NAIF Documentation

source

SPICE.lparse — Method.

lparse(list, delim, nmax)

Parse a list of items delimited by a single character.

Arguments

list: List of items delimited by delim

delim: Single character used to delimit items

nmax: Maximum number of items to return

Output

Returns an array with the items in the list, left justified.

References

NAIF Documentation

source

SPICE.lparsm — Method.

lparsm(list, delims, nmax)

Parse a list of items separated by multiple delimiters.

Arguments

list: List of items delimited by delim

delims: Single characters which delimit items

nmax: Maximum number of items to return

Output

Returns an array with the items in the list, left justified.

References

NAIF Documentation

source

SPICE.lparss — Method.

lparss(list, delims)

Parse a list of items separated by multiple delimiters, placing the resulting items into a set.

Arguments

list: List of items delimited by delim

delims: Single characters which delimit items

Output

Returns a set containing items in the list, left justified

References

NAIF Documentation

source

SPICE.lspcn — Method.

lspcn(body, et, abcorr)

Compute L_s, the planetocentric longitude of the sun, as seen from a specified body.

Arguments

body: Name of the central body

et: Epoch in seconds past J2000 TDB

abcorr: Aberration correction

Output

Returns the planetocentric longitude of the sun for the specified body at the specified time in radians.

References

NAIF Documentation

source

SPICE.lstle — Function.

lstle(x, array)

Given an element x and an array of non-decreasing elements (floats, integers, or strings), find the index of the largest array element less than or equal to x.

Arguments

x: Value to search against

arrays: Array of possible lower bounds

Output

Returns the index of the highest-indexed element in the input array that is less than or equal to x. The routine assumes the array elements are sorted in non-decreasing order.

If all elements of the input array are greater than x, the function returns 0.

References

NAIF Documentation

NAIF Documentation

NAIF Documentation

source

SPICE.lstlt — Function.

lstle(x, array)

Given an element x and an array of non-decreasing elements (floats, integers, or strings), find the index of the largest array element less than x.

Arguments

x: Value to search against

arrays: Array of possible lower bounds

Output

Returns the index of the highest-indexed element in the input array that is less than x. The routine assumes the array elements are sorted in non-decreasing order.

If all elements of the input array are greater than or equal to x, the function returns 0.

References

NAIF Documentation

NAIF Documentation

NAIF Documentation

source

SPICE.ltime — Method.

ltime(etobs, obs, dir, targ)

This routine computes the transmit (or receive) time of a signal at a specified target, given the receive (or transmit) time at a specified observer. The elapsed time between transmit and receive is also returned.

Arguments

etobs: Epoch of a signal at some observer

obs: NAIF ID of some observer

dir: Direction the signal travels ( "->" or "<-" )

targ: Time between transmit and receipt of the signal

Output

Returns the tuple (ettarg, elapsd).

ettarg: Epoch of the signal at the target

obs: NAIF ID of some observer

References

NAIF Documentation

source

SPICE.lx4dec — Method.

lx4dec(string, first)

Scan a string from a specified starting position for the end of a decimal number.

Arguments

string: Any character string

first: First character to scan from in string

Output

Returns the tuple (last, nchar).

last: Last character that is part of a decimal number. If there is no such character, last will be returned with the value first-1.

nchar: Number of characters in the decimal number

References

NAIF Documentation

source

SPICE.lx4num — Method.

lx4num(string, first)

Scan a string from a specified starting position for the end of a number.

Arguments

string: Any character string

first: First character to scan from in string

Output

Returns the tuple (last, nchar).

last: Last character that is part of a number. If there is no such character, last will be returned with the value first-1.

nchar: Number of characters in the number

References

NAIF Documentation

source

SPICE.lx4sgn — Method.

lx4sgn(string, first)

Scan a string from a specified starting position for the end of a signed integer.

Arguments

string: Any character string

first: First character to scan from in string

Output

Returns the tuple (last, nchar).

last: Last character that is part of a signed integer. If there is no such character, last will be returned with the value first-1.

nchar: Number of characters in the signed integer

References

NAIF Documentation

source

SPICE.lx4uns — Method.

lx4uns(string, first)

Scan a string from a specified starting position for the end of a unsigned integer.

Arguments

string: Any character string

first: First character to scan from in string

Output

Returns the tuple (last, nchar).

last: Last character that is part of an unsigned integer. If there is no such character, last will be returned with the value first-1.

nchar: Number of characters in the unsigned integer

References

NAIF Documentation

source

SPICE.m2eul — Method.

m2eul(r, axis3, axis2, axis1)

Factor a rotation matrix as a product of three rotations about specified coordinate axes.

Arguments

r: A rotation matrix to be factored

axis3: Number of the third rotation axis

axis2: Number of the second rotation axis

axis1: Number of the first rotation axis

Output

A tuple consisting of the third, second, and first Euler angles in radians.

References

NAIF Documentation

source

SPICE.m2q — Method.

m2q(r)

Find a unit quaternion corresponding to a specified rotation matrix.

Arguments

r: A rotation matrix

Output

A unit quaternion representing `r'

References

NAIF Documentation

source

SPICE.matchi — Method.

matchi(string, templ, wstr, wchar)

Determine whether a string is matched by a template containing wild cards. The pattern comparison is case-insensitive.

Arguments

string: String to be tested

templ: Template (with wild cards) to test against string

wstr: Wild string token

wchr: Wild character token

Output

Returns true if the string matches.

References

NAIF Documentation

source

SPICE.matchw — Method.

matchw(string, templ, wstr, wchar)

Determine whether a string is matched by a template containing wild cards.

Arguments

string: String to be tested

templ: Template (with wild cards) to test against string

wstr: Wild string token

wchr: Wild character token

Output

Returns true if the string matches.

References

NAIF Documentation

source

SPICE.maxd — Function.

maxd(args...)

Deprecated: Use max(args...) instead.

source

SPICE.maxi — Function.

maxi(args...)

Deprecated: Use max(args...) instead.

source

SPICE.mequ — Function.

mequ(m1, mout)

Deprecated: Use mout .= m1 instead.

source

SPICE.mequg — Function.

mequg(m1, mout)

Deprecated: Use mout .= m1 instead.

source

SPICE.mind — Function.

mind(args...)

Deprecated: Use min(args...) instead.

source

SPICE.mini — Function.

mini(args...)

Deprecated: Use min(args...) instead.

source

SPICE.mtxm — Function.

mtxm(m1, m2)

Deprecated: Use m1' * m2 instead.

source

SPICE.mtxmg — Function.

mtxmg(m1, m2)

Deprecated: Use m1' * m2 instead.

source

SPICE.mtxv — Function.

mtxv(m1,v2)

Deprecated: Use m1' * v2 instead.

source

SPICE.mtxvg — Function.

mtxvg(m1,v2)

Deprecated: Use m1' * v2 instead.

source

SPICE.mxm — Function.

mxm(m1, m2)

Deprecated: Use m1 * m2 instead.

source

SPICE.mxmg — Function.

mxmg(m1, m2)

Deprecated: Use m1 * m2 instead.

source

SPICE.mxmt — Function.

mxmt(m1, m2)

Deprecated: Use m1 * m2' instead.

source

SPICE.mxmtg — Function.

mxmtg(m1, m2)

Deprecated: Use m1 * m2' instead.

source

SPICE.mxv — Function.

mxv(m1,v2)

Deprecated: Use m1 * v2 instead.

source

SPICE.mxvg — Function.

mxvg(m1,v2)

Deprecated: Use m1 * v2 instead.

source

SPICE.namfrm — Method.

namfrm(frname)

Look up the frame ID code associated with a string.

Arguments

frname: The name of some reference frame

Output

The SPICE ID code of the frame.

References

NAIF Documentation

source

SPICE.ncpos — Method.

ncpos(str, chars, start)

Find the first occurrence in a string of a character NOT belonging to a collection of characters, starting at a specified location, searching forward.

Arguments

str: A string

chars: A collection of characters

start: Position to begin looking for a character not in chars

Output

Returns the index of the first character of str at or following index start that is not in the collection chars.

References

NAIF Documentation

source

SPICE.ncposr — Method.

ncposr(str, chars, start)

Find the first occurrence in a string of a character NOT belonging to a collection of characters, starting at a specified location, searching in reverse.

Arguments

str: A string

chars: A collection of characters

start: Position to begin looking for a character not in chars

Output

Returns the index of the last character of str at or before index start that is not in the collection chars.

References

NAIF Documentation

source

SPICE.nearpt — Method.

nearpt(positn, a, b, c)

This routine locates the point on the surface of an ellipsoid that is nearest to a specified position. It also returns the altitude of the position above the ellipsoid.

Arguments

positn: Position of a point in the bodyfixed frame

a: Length of semi-axis parallel to x-axis

b: Length of semi-axis parallel to y-axis

c: Length on semi-axis parallel to z-axis

Output

Returns a tuple consisting of npoint and alt.

npoint: Point on the ellipsoid closest to positn

alt: Altitude of positn above the ellipsoid

References

NAIF Documentation

source

SPICE.npedln — Method.

npedln(a, b, c, linept, linedr)

Find nearest point on a triaxial ellipsoid to a specified line, and the distance from the ellipsoid to the line.

Arguments

a: Length of semi-axis in the x direction

b: Length of semi-axis in the y direction

c: Length of semi-axis in the z direction

linept: Point on line

linedr: Direction vector of line

Output

Returns a tuple consisting of pnear and dist.

pnear: Nearest point on ellipsoid to line

dist: Distance of ellipsoid from line

References

NAIF Documentation

source

SPICE.npelpt — Method.

npelpt(point, ellips)

Find the nearest point on an ellipse to a specified point, both in three-dimensional space, and find the distance between the ellipse and the point.

Arguments

point: Point whose distance to an ellipse is to be found

ellips: A SPICE ellipse

Output

Returns a tuple consisting of pnear and dist.

pnear: Nearest point on ellipse to input point

dist: Distance of input point to ellipse

References

NAIF Documentation

source

SPICE.nplnpt — Method.

nplnpt(linept, linedr, point)

Find the nearest point on a line to a specified point, and find the distance between the two points.

Arguments

linept: Point on line

linedr: Direction vector of line

point: A second point

Output

Returns a tuple consisting of pnear and dist.

pnear: Nearest point on the line to point

dist: Distance between point and pnear

References

NAIF Documentation

source

SPICE.nvc2pl — Method.

nvc2pl(norm, point)

Make a SPICE plane from a normal vector and a point.

Arguments

norm: A normal vector...

constant: ...and a constant defining a plane

Output

Returns a struct representing the plane.

References

NAIF Documentation

source

SPICE.nvp2pl — Method.

nvp2pl(norm, point)

Make a SPICE plane from a normal vector and a point.

Arguments

norm: A normal vector...

point: ...and a point defining a plane

Output

Returns a struct representing the plane.

References

NAIF Documentation

source

SPICE.oscelt — Method.

oscelt(state, et, mu)

Determine the set of osculating conic orbital elements that corresponds to the state (position, velocity) of a body at some epoch.

Arguments

state: State of body at epoch of elements

et: Epoch of elements

mu: Gravitational parameter (GM) of primary body

Output

Returns the equivalent conic elements.

References

NAIF Documentation

source

SPICE.pdpool — Method.

pdpool(name, vals)

Insert double precision data into the kernel pool.

Arguments

name: The kernel pool name to associate with vals

vals: An array of values to insert into the kernel pool

References

NAIF Documentation

source

SPICE.pxform — Method.

Return the matrix that transforms position vectors from one specified frame to another at a specified epoch.

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/pxformc.html

source

SPICE.q2m — Method.

q2m(q...)

Find the rotation matrix corresponding to a specified unit quaternion.

Arguments

q: A unit quaternion (as any kind of iterable with four elements)

Output

A rotation matrix corresponding to q.

References

NAIF Documentation

source

SPICE.qdq2av — Method.

qdq2av(q,dq)

Derive angular velocity from a unit quaternion and its derivative with respect to time.

Arguments

q: Unit SPICE quaternion (as any kind of iterable with four elements)

dq: Derivative of `q' with respect to time

Output

Angular velocity vector defined by q' anddq'

References

NAIF Documentation

source

SPICE.qxq — Method.

qxq(q1,q2)

Multiply two quaternions.

Arguments

q1: First SPICE quaternion factor (as any kind of iterable with four elements)

q2: Second SPICE quaternion factor (as any kind of iterable with four elements)

Output

A quaternion corresponding to the product of q1' andq2'

References

NAIF Documentation

source

SPICE.rav2xf — Method.

rav2xf(rot, av)

Determine a state transformation matrix from a rotation matrix and the angular velocity of the rotation.

Arguments

rot: Rotation matrix

av: Angular velocity vector

Output

Returns state transformation matrix associated with rot and av.

References

NAIF Documentation

source

SPICE.recrad — Method.

recrad(rectan)

Convert rectangular coordinates to range, right ascension, and declination.

Arguments

rectan: Rectangular coordinates of a point

Output

Return the tuple (range, ra, dec).

range: Distance of the point from the origin

ra: Right ascension in radians

dec: Declination in radians

References

NAIF Documentation

source

SPICE.rotate — Method.

rotate(angle, iaxis)

Calculate the 3x3 rotation matrix generated by a rotation of a specified angle about a specified axis. This rotation is thought of as rotating the coordinate system.

Arguments

angle: Angle of rotation (radians)

iaxis: Axis of rotation (X=1, Y=2, Z=3)

Output

Returns rotation matrix associated with angle and iaxis.

References

NAIF Documentation

source

SPICE.spd — Method.

Returns the number of seconds in a day.

source

SPICE.spkcpo — Method.

Returns the state of a target body relative to a constant-position observer location.

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/spkcpoc.html

source

SPICE.spkezr — Method.

Returns the state of a target body relative to an observing body.

NAIF documentation

source

SPICE.spkpos — Method.

Returns the state of a target body relative to an observing body.

NAIF documentation

source

SPICE.swpool — Method.

swpool(agent, names)

Add a name to the list of agents to notify whenever a member of a list of kernel variables is updated.

Arguments

agent: The name of an agent to be notified after updates

names: Variable names whose update causes the notice

References

NAIF Documentation

source

SPICE.sxform — Method.

Return the state transformation matrix from one frame to another at a specified epoch.

https://naif.jpl.nasa.gov/pub/naif/toolkitdocs/C/cspice/sxformc.html

source

SPICE.timdef — Function.

timdef(action, item, value="")

Set and retrieve the defaults associated with calendar input strings.

Arguments

action: The kind of action to take, either :SET or :GET

item: The default item of interest. The items that may be requested are:

:CALENDAR with allowed values:

"GREGORIAN"

"JULIAN"

"MIXED"

:SYSTEM with allowed values:

"TDB"

"TDT"

"UTC"

:ZONE with allowed values (0 <= HR < 13 and 0 <= MN < 60):

"EST"

"EDT"

"CST"

"CDT"

"MST"

"MDT"

"PST"

"PDT"

"UTC+$HR"

"UTC-$HR"

"UTC+$HR:$MN"

"UTC-$HR:$MN"

Output

Returns the value associated with the default item.

References

NAIF Documentation

source

SPICE.timout — Function.

timout(et, pictur, lenout=128)

This routine converts an input epoch represented in TDB seconds past the TDB epoch of J2000 to a character string formatted to the specifications of a user's format picture.

Arguments

et: An epoch in seconds past the ephemeris epoch J2000

pictur: A format specification for the output string

lenout: The length of the output string plus 1 (default: 128)

Output

Returns a string representation of the input epoch.

References

NAIF Documentation

source

SPICE.tipbod — Method.

tipbod(ref, body, et)

Return a 3x3 matrix that transforms positions in inertial coordinates to positions in body-equator-and-prime-meridian coordinates.

Arguments

ref: Name of inertial reference frame to transform from

body: ID code of body

et: Epoch of transformation

Output

Returns transformation matrix from intertial position to prime meridian.

References

NAIF Documentation

source

SPICE.tisbod — Method.

tisbod(ref, body, et)

Return a 6x6 matrix that transforms states in inertial coordinates to states in body-equator-and-prime-meridian coordinates.

Arguments

ref: Name of inertial reference frame to transform from

body: ID code of body

et: Epoch of transformation

Output

Returns transformation matrix from intertial state to prime meridian.

References

NAIF Documentation

source

SPICE.tkvrsn — Function.

tkvrsn(item=:TOOLKIT)

Given an item such as the Toolkit or an entry point name, return the latest version string.

References

NAIF Documentation

source

SPICE.tparse — Method.

tparse(string)

Parse a time string and return seconds past the J2000 epoch on a formal calendar.

Arguments

string: Input time string in UTC

Output

Returns UTC expressed in seconds since J2000.

References

NAIF Documentation

source

SPICE.tpictr — Function.

tpictr(sample, lenout=80)

Given a sample time string, create a time format picture suitable for use by the routine timout.

Arguments

sample: A sample time string

lenout: The length for the output picture string (default: 80)

Output

Returns a format picture that describes sample.

References

NAIF Documentation

source

SPICE.trace — Function.

trace(matrix)

Deprecated: Use LinearAlgebra.tr(matrix) instead.

source

SPICE.tsetyr — Method.

tsetyr(year)

Set the lower bound on the 100 year range.

Arguments

-year: Lower bound on the 100 year interval of expansion

References

NAIF Documentation

source

SPICE.twopi — Function.

twopi()

Deprecated: Use 2π instead.

source

SPICE.twovec — Method.

twovec(axdef, indexa, plndef, indexp)

Find the transformation to the right-handed frame having a given vector as a specified axis and having a second given vector lying in a specified coordinate plane.

Arguments

axdef: Vector defining a principal axis

indexa: Principal axis number of axdef (X=1, Y=2, Z=3)

plndef: Vector defining (with axdef) a principal plane

indexp: Second axis number (with indexa) of principal plane

Output

Returns output rotation matrix.

References

NAIF Documentation

source

SPICE.tyear — Method.

tyear()

Returns the number of seconds per tropical year.

References

NAIF Documentation

source

SPICE.ucase — Function.

ucase(in)

Deprecated: Use uppercase(in) instead.

source

SPICE.ucrss — Function.

ucrss(v1, v2)

Deprecated: Use LinearAlgebra.normalize(LinearAlgebra.cross(v1, v2)) instead.

source

SPICE.uddf — Method.

uddf(udfunc, x, dx)

Routine to calculate the first derivative of a caller-specified function using a three-point estimation.

Arguments

udfunc: A callable that computes the scalar value of interest, e.g. f(x::Float64) -> Float64.

x: Independent variable of 'udfunc'

dx: Interval from x for derivative calculation

Output

Returns the approximate derivative of udfunc at x.

References

NAIF Documentation

source

SPICE.unitim — Method.

unitim(epoch, insys, outsys)

Transform time from one uniform scale to another.

Arguments

epoch: An epoch to be converted

insys: The time scale associated with the input epoch

outsys: The time scale associated with the function value

The uniform time scales are:

:TAI

:TDT

:TDB

:ET

:JED

:JDTDB

:JDTDT

Output

Returns the time in the system specified by outsys that is equivalent to the epoch in the insys time scale.

References

NAIF Documentation

source

SPICE.unload — Method.

unload(file)

Unload a SPICE kernel.

Arguments

file: The file name of a kernel to unload

References

NAIF Documentation

source

SPICE.unorm — Function.

unorm(v1)

Deprecated: Use (LinearAlgebra.normalize(v1), LinearAlgebra.norm(v1)) instead.

source

SPICE.unormg — Function.

unormg(v1)

Deprecated: Use (LinearAlgebra.normalize(v1), LinearAlgebra.norm(v1)) instead.

source

SPICE.utc2et — Method.

utc2et(utcstr)

Convert an input time from Calendar or Julian Date format, UTC, to ephemeris seconds past J2000.

Arguments

utcstr: Input time string, UTC

Output

Returns the equivalent of utcstr, expressed in ephemeris seconds past J2000.

References

NAIF Documentation

source

SPICE.vadd — Function.

vadd(v1, v2)

Deprecated: Use v1 .+ v2 instead.

source

SPICE.vaddg — Function.

vaddg(v1, v2)

Deprecated: Use v1 .+ v2 instead.

source

SPICE.valid! — Method.

valid!(set::SpiceCell{T}) where T

Create a valid SPICE set from a SPICE Cell of any data type.

Arguments

set: Set to be validated

Output

Returns the validated set with ordered elements and duplicates removed.

References

NAIF Documentation

source

SPICE.vcrss — Function.

vcrss(v1, v2)

Deprecated: Use LinearAlgebra.cross(v1, v2) instead.

source

SPICE.vdist — Function.

vdist(v1, v2)

Deprecated: Use LinearAlgebra.norm(v1 .- v2) instead.

source

SPICE.vdistg — Function.

vdistg(v1, v2)

Deprecated: Use LinearAlgebra.norm(v1 .- v2) instead.

source

SPICE.vdot — Function.

vdot(v1, v2)

Deprecated: Use LinearAlgebra.dot(v1, v2) instead.

source

SPICE.vdotg — Function.

vdotg(v1, v2)

Deprecated: Use LinearAlgebra.dot(v1, v2) instead.

source

SPICE.vequ — Function.

vequ(v1, v2)

Deprecated: Use v1 .= v2 instead.

source

SPICE.vequg — Function.

vequg(v1, v2)

Deprecated: Use v1 .= v2 instead.

source

SPICE.vhat — Function.

vhat(v1)

Deprecated: Use LinearAlgebra.normalize(v1) instead.

source

SPICE.vhatg — Function.

vhatg(v1)

Deprecated: Use LinearAlgebra.normalize(v1) instead.

source

SPICE.vlcom — Function.

vlcom(a, v1, b, v2)

Deprecated: Use a .* v1 .+ b .* v2 instead.

source

SPICE.vlcom3 — Function.

vlcom3(a, v1, b, v2, c, v3)

Deprecated: Use a .* v1 .+ b .* v2 .+ c .* v3 instead.

source

SPICE.vlcomg — Function.

vlcomg(a, v1, b, v2)

Deprecated: Use a .* v1 .+ b .* v2 instead.

source

SPICE.vminug — Function.

vminug(vin)

Deprecated: Use -vin instead.

source

SPICE.vminus — Function.

vminus(vin)

Deprecated: Use -vin instead.

source

SPICE.vnorm — Function.

vnorm(v1)

Deprecated: Use LinearAlgebra.norm(v1) instead.

source

SPICE.vnormg — Function.

vnormg(v1, v2)

Deprecated: Use LinearAlgebra.norm(v1) instead.

source

SPICE.vpack — Function.

vpack(x, y, z)

Deprecated: Use [x, y, z] instead.

source

SPICE.vperp — Method.

vperp(a, b)

Find the component of a vector that is perpendicular to a second vector.

Arguments

a: The vector whose orthogonal component is sought

b: The vector used as the orthogonal reference

Output

Returns the component a orthogonal to b.

References

NAIF Documentation

source

SPICE.vprjp — Method.

vprjp(vin, plane)

Project a vector onto a specified plane, orthogonally.

Arguments

vin: Vector to be projected

plane: Plane onto which vin is projected

Output

Returns the vector resulting from the projection.

References

NAIF Documentation

source

SPICE.vprjpi — Method.

vprjpi(vin, projpl, invpl)

Find the vector in a specified plane that maps to a specified vector in another plane under orthogonal projection.

Arguments

vin: The projected vector

projpl: Plane containing vin

invpl: Plane containing inverse image of vin

Output

Returns the inverse projection of vin or nothing if vin could not be calculated.

References

NAIF Documentation

source

SPICE.vproj — Method.

vproj(a, b)

Finds the projection of one vector onto another vector. All vectors are 3-dimensional.

Arguments

a: The vector to be projected

b: The vector onto which a is to be projected

Output

Returns the projection of a onto b.

References

NAIF Documentation

source

SPICE.vrel — Method.

vrel(v1, v2)

Return the relative difference between two 3-dimensional vectors.

Arguments

v1, v2: Two three-dimensional input vectors

Output

Returns the relative differences between v1 and v2.

References

NAIF Documentation

source

SPICE.vrelg — Method.

vrelg(v1, v2)

Return the relative difference between two vectors.

Arguments

v1, v2: Input vectors

Output

Returns the relative differences between v1 and v2.

References

NAIF Documentation

source

SPICE.vrotv — Method.

vrotv(v, axis, theta)

Rotate a vector about a specified axis vector by a specified angle and return the rotated vector.

Arguments

v: Vector to be rotated

axis: Axis of the rotation

theta: Angle of rotation (radians)

Output

Result of rotating v about axis by theta.

References

NAIF Documentation

source

SPICE.vscl — Function.

vscl(s, v1)

Deprecated: Use s .* v1 instead.

source

SPICE.vsclg — Function.

vsclg(s, v1)

Deprecated: Use s .* v1 instead.

source

SPICE.vsep — Method.

vsep(v1, v2)

Return the sepative difference between two 3-dimensional vectors.

Arguments

v1, v2: Two three-dimensional input vectors

Output

Returns the angle between v1 and v2 in radians.

References

NAIF Documentation

source

SPICE.vsepg — Method.

vsepg(v1, v2)

Return the sepative difference between two vectors.

Arguments

v1, v2: Input vectors

Output

Returns the angle between v1 and v2 in radians.

References

NAIF Documentation

source

SPICE.vsub — Function.

vsub(v1, v2)

Deprecated: Use v1 .- v2 instead.

source

SPICE.vsubg — Function.

vsubg(v1, v2)

Deprecated: Use v1 .- v2 instead.

source

SPICE.vtmv — Function.

vtmv(v1, matrix, v2)

Deprecated: Use v1' * matrix * v2 instead.

source

SPICE.vtmvg — Function.

vtmvg(v1, matrix, v2)

Deprecated: Use v1' * matrix * v2 instead.

source

SPICE.vupack — Function.

vupack(v)

Deprecated: Use x, y, z = v instead.

source

SPICE.vzero — Function.

vzero(v1)

Deprecated: Use iszero(v1) instead.

source

SPICE.vzerog — Function.

vzerog(v1, v2)

Deprecated: Use iszero(v1) instead.

source

SPICE.wncard — Method.

wncard(window)

Return the cardinality (number of intervals) of a double precision window.

Arguments

window: Input window

References

NAIF Documentation

source

SPICE.wncomd — Method.

wncomd(window, left, right)

Determine the complement of a double precision window with respect to a specified interval.

Arguments

window: Input window.

left: Left endpoint of the complement interval

right: Right endpoint of the complement interval

Output

Returns the complement of window with respect to [left,right].

References

NAIF Documentation

source

SPICE.wncond! — Method.

wncond!(window, left, right)

Contract each of the intervals of a double precision window.

Arguments

window: Window to be contracted

left: Amount added to each left endpoint

right: Amount subtracted from each right endpoint

Output

Returns the contracted window.

References

NAIF Documentation

source

SPICE.wndifd — Method.

wndifd(a, b)

Place the difference of two double precision windows into a third window.

Arguments

a: Input window

b: Input window

Output

Returns a window containing the difference of a and b.

References

NAIF Documentation

source

SPICE.wnelmd — Method.

wnelmd(window, point)

Determine whether a point is an element of a double precision window.

Arguments

window: Input window

point: Input point

Output

Returns true if point is an element of window.

References

NAIF Documentation

source

SPICE.wnexpd! — Method.

wnexpd(window, left, right)

Expand each of the intervals of a double precision window.

Arguments

left: Amount subtracted from each left endpoint

right: Amount added to each right endpoint

Output

Returns the expanded window.

References

NAIF Documentation

source

SPICE.wnextd! — Method.

wnextd!(window, side)

Extract the left or right endpoints from a double precision window.

Arguments

window: Window to be extracted

side: Extract left (:L) or right (:R) endpoints

Output

Returns the extracted window.

References

NAIF Documentation

source

SPICE.wnfetd — Method.

wnfetd(window, n)

Fetch a particular interval from a double precision window.

Arguments

window: Input window

n: Index of interval to be fetched

Output

Returns a tuple consisting of the left and right endpoints of the n-th interval in the input window.

References

NAIF Documentation

source

SPICE.wnfild! — Method.

wnfild!(window, small)

Fill small gaps between adjacent intervals of a double precision window.

Arguments

window: Window to be filled

small: Limiting measure of small gaps

Output

Returns the updated window.

References

NAIF Documentation

source

SPICE.wnfltd! — Method.

wnfild!(window, small)

Filter (remove) small intervals from a double precision window.

Arguments

window: Window to be filtered

small: Limiting measure of small intervals

Output

Returns the updated window.

References

NAIF Documentation

source

SPICE.wnincd — Method.

wnincd(window, left, right)

Determine whether an interval is included in a double precision window.

Arguments

window: Input window

left: Left endpoint of the input interval

right: Right endpoint of the input interval

Output

Returns true when (left, right) is contained in window.

References

NAIF Documentation

source

SPICE.wninsd! — Method.

wninsd!(window, left, right)

Insert an interval into a double precision window.

Arguments

window: Input window

left: Left endpoint of the new interval

right: Right endpoint of the new interval

Output

Returns the updated windows.

References

NAIF Documentation

source

SPICE.wnintd — Method.

wnintd(a, b)

Place the intersection of two double precision windows into a third window.

Arguments

a: Input window

b: Input window

Output

Returns a window containing the intersection of a and b.

References

NAIF Documentation

source

SPICE.wnreld — Method.

wnreld(a, op, b)

Compare two double precision windows.

Note

Consider using overloaded operators instead, i.e. a == b, a ⊆ b, and a ⊊ b.

Arguments

a: First window

op: Comparison operator

b: Second window

Output

Returns the result of comparison a (op) b.

References

NAIF Documentation

source

SPICE.wnsumd — Method.

wnsumd(window)

Summarize the contents of a double precision window.

Arguments

window: Window to be summarized

Output

Returns a tuple consisting of:

meas: Total measure of intervals in window

avg: Average measure

stddev: Standard deviation

shortest: Location of shortest interval

longest: Location of longest interval

References

NAIF Documentation

source

SPICE.wnunid — Method.

wnunid(a, b)

Place the union of two double precision windows into a third window.

Arguments

a: Input window

b: Input window

Output

Returns a window containing the union of a and b.

References

NAIF Documentation

source

SPICE.wnvald! — Method.

wnvald!(window)

Form a valid double precision window from the contents of a window array.

Arguments

window: A (possibly uninitialized) SpiceDoubleCell containing endpoints of (possibly unordered and non-disjoint) intervals.

Output

Returns the validated window.

References

NAIF Documentation

source

SPICE.xf2eul — Method.

xf2eul(xform, axisa, axisb, axisc)

Convert a state transformation matrix to Euler angles and their derivatives with respect to a specified set of axes.

Arguments

xform: A state transformation matrix

axisa: Axis A of the Euler angle factorization

axisb: Axis B of the Euler angle factorization

axisc: Axis C of the Euler angle factorization

Output

Returns a tuple of an array of Euler angles and their derivatives and a boolean that indicates whether these are a unique representation.

References

NAIF Documentation

source

SPICE.xf2rav — Method.

xf2rav(xform)

Determines the rotation matrix and angular velocity of the rotation from a state transformation matrix.

Arguments

xform: State transformation matrix

Output

Returns a tuple of the rotation matrix and the angular velocity vector associated with xform.

References

NAIF Documentation

source

SPICE.xfmsta — Method.

xfmsta(input_state, input_coord_sys, output_coord_sys, body)

Transform a state between coordinate systems.

Arguments

input_state: Input state

input_coord_sys: Current (input) coordinate system

`outputcoordsys: Desired (output) coordinate system

body: Name or NAIF ID of body with which coordinates are associated (if applicable)

Output

Returns the converted output state.

References

NAIF Documentation

source

SPICE.xpose — Function.

xpose(matrix)

Deprecated: Use transpose(matrix) instead.

source

SPICE.xpose6 — Function.

xpose6(matrix)

Deprecated: Use transpose(matrix) instead.

source

SPICE.xposeg — Function.

xposeg(matrix)

Deprecated: Use transpose(matrix) instead.

source